requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Homie3==0.3.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.3
netifaces==0.10.9
packaging==20.1
paho-mqtt==1.5.0
pluggy==0.13.1
ptvsd==4.3.2
py==1.8.1
pydantic==1.4
pylint==2.4.4
pytest==5.3.5
python-dotenv==0.12.0
rope==0.16.0
six==1.14.0
typed-ast==1.4.1
wcwidth==0.1.8
wincertstore==0.2
wrapt==1.11.2
yarl==1.4.2
zipp==2.2.0
